implement caching grid in api.ts + setVizStore.ts
Add routine to check viewport bbox for intersections with one
or more Bbox from cacheGrid.ts, and make one or more requests for
this cacheable Bbox rather than the viewport bbox. Intention is
to quantise requsts made to the api to take advantage of its cache.

Routine makes use of shallow (one request deep) local cache to
avoid requesting identical data repeatedly (this is an issue with
the cache grid, as the cachable Bboxes are often larger than the
viewport is, and so moving the viewport may result in identical
geodata api requests).