swap jest for vitest (temp while issue with import.meta.env fixed)
Use vitest instead of jest as test framework (minor refactoring
needed for vitest global injection and use of mocks). This
change needed to test code that makes use of pseudo-env vars
with the import.meta.env syntax (jest does not like this).
This should be fixable when we figure out the issues with using
.env files, so should be able to go back to jest at some point.